// src/schemas/admin/auth/login.schema.ts
import { z } from "zod";

export const loginSchema = z.object({
	email: z.string().min(1, "Email é obrigatório").email("Email inválido").trim(),
	password: z.string().min(1, "Senha é obrigatória").trim(),
	rememberMe: z.boolean().optional(),
});

export type LoginSchema = z.infer<typeof loginSchema>;